benchmarks-wob/full-reduced/sentiment-analysis/sentiment-full-wob/requirements.txt
parliament-functions==0.1.0
requests
redis
pymongo
